# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plataforma de Atendimento Omnichannel (SaaS Multi-Tenant) API",
    "description": "Especificação completa da API Go (Fiber + PostgreSQL RLS + Redis) estilo Chatwoot + Kommo + Flow Engine",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/campaigns": {
      "get": {
        "summary": "Lista campanhas de disparo em massa",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Lista de campanhas" } }
      }
    },
...
    "/campaigns/{id}/import-csv": {
      "post": {
        "summary": "Importa CSV de contatos para campanha",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Importado" } }
      }
    },
...
  }
}